+++ DelvUI/changelog.md
@@ -7,6 +7,7 @@ Features:
 - Added alpha bar to color selections.
 - Making party frames and status effects list areas more clear now. When you open DelvUI's settings window, these areas will automatically show reflecting the real size of the hud elements.
 - Pressing Escape will now close DelvUI's setting window.
+- Added sub-option to Hide DelvUI outside of combat to show when weapons are drawn.
 
 Fixes:
 - Fixed clicks not working on game windows when they are on top of unit frames or party frames.